new pytorch with gpu requirements + removed "Unsure" directory from repo
