+++ .github/workflows/optimised_integration.yaml
@@ -37,7 +37,7 @@ jobs:
       - name: Install optimised dependencies
         run: |
           python -m pip install -r requirements-optimised.txt
-          python -m pip install .[optimised]
+          python -m pip install .
 
       - name: Run optimised unit tests
         run: pytest -m 'not slow and not integration and not pytorch'
