--- coppafish/Unsure/register/__init__.py
@@ -1,3 +1,0 @@
-from .base import get_transform, get_average_transform, icp, transform_from_scale_shift, get_single_affine_transform, \
-    transform_from_round_scale_shift
-from .check_transforms import check_transforms
