+++ coppafish/__init__.py
@@ -1,5 +1,4 @@
 from . import call_spots, extract, find_spots, omp, pipeline, setup, spot_colors, stitch, utils, robominnie
-from .Unsure import register
 from .pipeline.run import run_pipeline
 from .setup import Notebook, NotebookPage
 from .utils.pciseq import export_to_pciseq
