+++ coppafish/call_spots/background_pytorchgpu.py
@@ -30,18 +30,20 @@ def fit_background(
         - background_vectors `float [n_channels x n_rounds x n_channels]`.
             background_vectors[c] is the background vector for channel c.
     """
+    device = spot_colors.device
     # Ensure weight_shift > 1e-20 to avoid blow up to infinity.
     weight_shift = np.clip(weight_shift, 1e-20, torch.inf)
 
     n_rounds, n_channels = spot_colors[0].shape
-    background_vectors = torch.repeat_interleave(torch.eye(n_channels)[:, None, :], n_rounds, dim=1)
+    background_vectors = torch.repeat_interleave(torch.eye(n_channels, device=device)[:, None, :], n_rounds, dim=1)
     # give background_vectors an L2 norm of 1 so can compare coefficients with other genes.
     background_vectors = background_vectors / torch.linalg.norm(background_vectors, axis=(1, 2), keepdims=True)
 
     weight_factor = 1 / (torch.abs(spot_colors) + weight_shift)
     spot_weight = spot_colors * weight_factor
-    background_weight = torch.ones((1, n_rounds, n_channels)) * background_vectors[0, 0, 0] * weight_factor
+    background_weight = torch.ones((1, n_rounds, n_channels), device=device) * background_vectors[0, 0, 0] * weight_factor
     coef = torch.sum(spot_weight * background_weight, dim=1) / torch.sum(background_weight ** 2, dim=1)
-    residual = spot_colors - coef[:, None] * torch.ones((1, n_rounds, n_channels)) * background_vectors[0, 0, 0]
+    residual = spot_colors \
+        - coef[:, None] * torch.ones((1, n_rounds, n_channels), device=device) * background_vectors[0, 0, 0]
 
     return residual.type(torch.float32), coef.type(torch.float32), background_vectors.type(torch.float32)
