+++ coppafish/robominnie/test/test_integration.py
@@ -216,6 +216,6 @@ def test_pdf_builder() -> None:
 
 
 if __name__ == "__main__":
-    test_integration_small_two_tile()
+    test_integration_004()
     test_pdf_builder()
     test_viewers()
