+++ pyproject.toml
@@ -2,6 +2,7 @@
 markers = [
     "slow: marks tests as slow (deselect with '-m \"not slow\"')",
     "integration: integration tests",
-    "optimised: tests with this mark only work with optimised packages installed",
-    "pytorch: tests requiring the pytorch installation option"
+    "optimised: tests requiring requirements-optimised.txt packages",
+    "pytorch: tests requiring requirements-pytorch.txt packages"
+    "pytorchgpu: tests requiring requirements-pytorchgpu.txt packages and an available cuda device"
 ]
